+++ joulescope_ui/command_processor.py
@@ -25,14 +25,14 @@ Provide centralized application state management include:
 
 Alternatives and references include:
 
-* [QSettings](https://doc.qt.io/qt-5/qsettings.html) -
-  [PySide](https://doc.qt.io/qtforpython/PySide2/QtCore/QSettings.html)
+* [QSettings](https://doc.qt.io/qt-6/qsettings.html) -
+  [PySide](https://doc.qt.io/qtforpython/PySide6/QtCore/QSettings.html)
 * [Implementing a preferences dialog window in PyQt](https://stackoverflow.com/questions/39023584/implementing-a-preferences-dialog-window-in-pyqt)
 * [PyQtConfig](https://www.mfitzp.com/article/pyqtconfig/)
 
 """
 
-from PySide2 import QtCore
+from PySide6 import QtCore
 import logging
 import weakref
 import threading