+++ joulescope_ui/main.py
@@ -15,13 +15,16 @@
 # https://stackoverflow.com/questions/11874767/real-time-plotting-in-while-loop-with-matplotlib
 # https://wiki.qt.io/Gallery_of_Qt_CSS_Based_Styles
 
+from PySide6 import QtCore, QtGui, QtWidgets
+import pyqtgraph
+pyqtgraph.setConfigOptions(useOpenGL=True)
+
 import appnope
 import os
 import platform
 import sys
 from . import __version__
 import joulescope
-from PySide2 import QtCore, QtGui, QtWidgets
 from joulescope_ui.error_window import Ui_ErrorWindow
 from joulescope_ui.widgets import widget_register
 from joulescope.usb import DeviceNotify
@@ -56,6 +59,7 @@ import time
 import threading
 import webbrowser
 import logging
+
 log = logging.getLogger(__name__)
 
 
@@ -417,7 +421,7 @@ class MainWindow(QtWidgets.QMainWindow):
                             detail='This command uses General/_mru_open to undo.')
 
         # Device selection
-        self.device_action_group = QtWidgets.QActionGroup(self)
+        self.device_action_group = QtGui.QActionGroup(self)
         self._device_disable = DeviceDisable()
         self._device_add(self._device_disable)
 
@@ -444,11 +448,11 @@ class MainWindow(QtWidgets.QMainWindow):
         # Add global keyboard shortcuts for the application
         # Attempted app.installEventFilter(self) with def eventFilter
         # but need to handle gets ShortcutOverride, KeyPress, KeyRelease, multiple times
-        self._shortcut_undo = QtWidgets.QShortcut(QtGui.QKeySequence.Undo, self)
+        self._shortcut_undo = QtGui.QShortcut(QtGui.QKeySequence.Undo, self)
         self._shortcut_undo.activated.connect(lambda: self._cmdp.invoke('!undo'))
-        self._shortcut_redo = QtWidgets.QShortcut(QtGui.QKeySequence.Redo, self)
+        self._shortcut_redo = QtGui.QShortcut(QtGui.QKeySequence.Redo, self)
         self._shortcut_redo.activated.connect(lambda: self._cmdp.invoke('!redo'))
-        self._shortcut_spacebar = QtWidgets.QShortcut(QtGui.QKeySequence(QtCore.Qt.Key_Space), self)
+        self._shortcut_spacebar = QtGui.QShortcut(QtGui.QKeySequence(QtCore.Qt.Key_Space), self)
         self._shortcut_spacebar.activated.connect(self._on_spacebar)
 
         if not self._cmdp.restore_success:
@@ -465,7 +469,7 @@ class MainWindow(QtWidgets.QMainWindow):
                 w = self._menu_setup(value, wroot)
                 w['__root__'] = wroot
             else:
-                w = QtWidgets.QAction(parent)
+                w = QtGui.QAction(parent)
                 w.setText(name)
                 if callable(value):
                     w.triggered.connect(value)
@@ -521,7 +525,7 @@ class MainWindow(QtWidgets.QMainWindow):
         menu.clear()
 
         developer = self._cmdp['General/developer']
-        self._profile_action_group = QtWidgets.QActionGroup(menu)
+        self._profile_action_group = QtGui.QActionGroup(menu)
         self._profile_action_group.setExclusive(True)
         for profile in sorted(self._cmdp.preferences.profiles):
             if profile == 'defaults':
@@ -999,7 +1003,7 @@ class MainWindow(QtWidgets.QMainWindow):
     def _device_add(self, device):
         """Add device to the user interface"""
         log.info('_device_change add %s', device)
-        action = QtWidgets.QAction(str(device), self)
+        action = QtGui.QAction(str(device), self)
         action.setCheckable(True)
         action.setChecked(False)
         action.triggered.connect(lambda x: self._cmdp.invoke('!Device/open', device))
@@ -1106,7 +1110,7 @@ class MainWindow(QtWidgets.QMainWindow):
         dialog.setRange(0, 1000)
         dialog.setCancelButtonText('Cancel')
         dialog.setWindowModality(QtCore.Qt.WindowModal)
-        width = QtGui.QFontMetrics(dialog.font()).width(text) + 100
+        width = QtGui.QFontMetrics(dialog.font()).horizontalAdvance(text) + 100
         dialog.resize(width, dialog.height())
         self._progress_dialog = dialog
         return dialog
@@ -1903,15 +1907,6 @@ def run(device_name=None, log_level=None, file_log_level=None, filename=None,
 
         log.info('Arguments: %s', (sys.argv, ))
         log.info('Start Qt')
-        try:
-            log.info('Configure high DPI scaling')
-            # http://doc.qt.io/qt-5/highdpi.html
-            # https://vicrucann.github.io/tutorials/osg-qt-high-dpi/
-            if sys.platform.startswith('win'):
-                ctypes.windll.user32.SetProcessDPIAware()
-            QtWidgets.QApplication.setAttribute(QtCore.Qt.AA_EnableHighDpiScaling)
-        except Exception:
-            log.exception('while configuring high DPI scaling')
         app = QtWidgets.QApplication(sys.argv)
         resource_list = [
             ('joulescope_ui', 'resources.rcc'),