+++ joulescope_ui/themes/manager.py
@@ -19,7 +19,7 @@ import json
 import logging
 import os
 import re
-from PySide2 import QtCore
+from PySide6 import QtCore
 
 
 log = logging.getLogger(__name__)