+++ joulescope_ui/widgets/gpio/gpio_widget.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from PySide2 import QtCore, QtWidgets
+from PySide6 import QtCore, QtWidgets
 from .gpio_widget_ui import Ui_GpioWidget
 from joulescope_ui.preferences import options_enum, to_bool
 from joulescope_ui.ui_util import comboBoxConfig, comboBoxSelectItemByText