+++ joulescope_ui/widgets/single_value/single_value_widget.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from PySide2 import QtCore, QtGui, QtWidgets
+from PySide6 import QtCore, QtGui, QtWidgets
 import math
 from joulescope.units import unit_prefix
 from joulescope_ui.units import convert_units